refactor(json): separate cold codec construction
