+++ java/fory-json/src/main/java/org/apache/fory/json/codegen/JsonWriterCodegen.java
@@ -24,46 +24,32 @@ import static org.apache.fory.codegen.ExpressionUtils.cast;
 import static org.apache.fory.codegen.ExpressionUtils.eq;
 import static org.apache.fory.codegen.ExpressionUtils.inline;
 
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.UUID;
 import org.apache.fory.codegen.Code;
 import org.apache.fory.codegen.CodegenContext;
 import org.apache.fory.codegen.Expression;
 import org.apache.fory.codegen.Expression.Reference;
 import org.apache.fory.codegen.ExpressionOptimizer;
 import org.apache.fory.json.ForyJsonException;
-import org.apache.fory.json.codec.ArrayCodec;
 import org.apache.fory.json.codec.ObjectCodec;
 import org.apache.fory.json.codec.ObjectCodec.AnyInfo;
-import org.apache.fory.json.codec.ScalarCodecs;
-import org.apache.fory.json.codec.StringObjectWriter;
-import org.apache.fory.json.codec.StringWriterCodec;
-import org.apache.fory.json.codec.Utf8ObjectWriter;
-import org.apache.fory.json.codec.Utf8WriterCodec;
 import org.apache.fory.json.meta.JsonFieldInfo;
 import org.apache.fory.json.meta.JsonFieldKind;
-import org.apache.fory.json.writer.StringJsonWriter;
-import org.apache.fory.json.writer.Utf8JsonWriter;
 import org.apache.fory.reflect.TypeRef;
 
 /**
  * Shared generation mechanics for concrete String and UTF-8 object-writer capabilities.
  *
- * <p>The nested generators own representation-specific field prefixes, scalar stores, and child
+ * <p>The concrete generators own representation-specific field prefixes, scalar stores, and child
  * capability types. This base shares source-construction algorithms only after the concrete writer
  * is selected; it is not a runtime output mode. Generated writers retain precomputed field tokens
  * and concrete child capabilities, fuse safe object prefixes, and split wide objects into bounded
  * methods to protect compiler and inlining budgets without adding per-field dispatch.
  */
 abstract class JsonWriterCodegen {
-  private static final int MIN_STRING_SPLIT_MEMBERS = 10;
-  private static final int MIN_UTF8_SPLIT_MEMBERS = 12;
   // Bound field logic in independently compiled generated methods. The entry method keeps a small
   // tail inline below; wide objects use fewer bounded calls without adding runtime dispatch.
   private static final int MAX_MEMBERS_PER_METHOD = 16;
@@ -151,7 +137,7 @@ abstract class JsonWriterCodegen {
     return JsonCodegen.usesWriteCodec(property);
   }
 
-  private static Reference fieldRef(String name, Class<?> type) {
+  static Reference fieldRef(String name, Class<?> type) {
     return Reference.fieldRef(name, TypeRef.of(type));
   }
 
@@ -490,99 +476,17 @@ abstract class JsonWriterCodegen {
     return expressions;
   }
 
-  final PrefixFields stringPrefixFields(JsonFieldInfo[] properties, boolean objectStartFused) {
-    PrefixFields fields = new PrefixFields(properties.length);
-    boolean commaKnown = objectStartFused;
-    for (int i = 0; i < properties.length; i++) {
-      JsonFieldInfo property = properties[i];
-      if (usesPrefix(property)) {
-        if (objectStartFused && i == 0) {
-          if (!canPackStringUtf16Prefix(property, false)) {
-            fields.name[i] = true;
-          }
-        } else {
-          markStringUtf16PrefixField(property, commaKnown, fields, i);
-        }
-      }
-      if (property.writeNull()) {
-        commaKnown = true;
-      }
-    }
-    return fields;
-  }
+  static final class PrefixFields {
+    final boolean[] name;
+    final boolean[] comma;
 
-  private static void markStringUtf16PrefixField(
-      JsonFieldInfo property, boolean commaKnown, PrefixFields fields, int id) {
-    if (!commaKnown) {
-      fields.name[id] = true;
-      fields.comma[id] = true;
-      return;
-    }
-    if (!canPackStringUtf16Prefix(property, true)) {
-      fields.comma[id] = true;
-    }
-  }
-
-  final PrefixFields utf8PrefixFields(JsonFieldInfo[] properties, boolean objectStartFused) {
-    PrefixFields fields = new PrefixFields(properties.length);
-    boolean commaKnown = objectStartFused;
-    for (int i = 0; i < properties.length; i++) {
-      JsonFieldInfo property = properties[i];
-      if (usesPrefix(property)) {
-        if (i == 0
-            && !objectStartFused
-            && !property.writeNull()
-            && Utf8Generator.canPackObjectStartString(property)) {
-          // The generated first-field branch consumes neither ordinary prefix field.
-        } else if (objectStartFused && i == 0) {
-          if (!canPackPrefix(property, false)) {
-            fields.name[i] = true;
-          }
-        } else if (!commaKnown) {
-          if (!canUsePackedDynamicPrefix(property)
-              || !canPackSinglePrefix(property, false)
-              || !canPackSinglePrefix(property, true)) {
-            fields.name[i] = true;
-            fields.comma[i] = true;
-          }
-        } else if (!canPackPrefix(property, true)) {
-          fields.comma[i] = true;
-        }
-      }
-      if (property.writeNull()) {
-        commaKnown = true;
-      }
-    }
-    return fields;
-  }
-
-  private boolean canUsePackedDynamicPrefix(JsonFieldInfo property) {
-    if (property.writeNull() && !property.writeRawType().isPrimitive()) {
-      return false;
-    }
-    switch (property.writeKind()) {
-      case BYTE:
-      case SHORT:
-      case INT:
-      case LONG:
-      case STRING:
-        return true;
-      default:
-        return false;
-    }
-  }
-
-  private static final class PrefixFields {
-    private final boolean[] name;
-    private final boolean[] comma;
-
-    private PrefixFields(int size) {
+    PrefixFields(int size) {
       name = new boolean[size];
       comma = new boolean[size];
     }
   }
 
-  private boolean usesPrefix(JsonFieldInfo property) {
+  final boolean usesPrefix(JsonFieldInfo property) {
     JsonFieldKind kind = property.writeKind();
     return kind != JsonFieldKind.BOOLEAN && kind != JsonFieldKind.ENUM
         || property.writeNull() && !property.writeRawType().isPrimitive();
@@ -1096,95 +1000,13 @@ abstract class JsonWriterCodegen {
     return expressions;
   }
 
-  private static Expression[] packedPrefixArgs(
-      JsonFieldInfo property, boolean comma, Expression... extraArgs) {
-    byte[] prefix = comma ? property.utf8CommaNamePrefix() : property.utf8NamePrefix();
-    Expression[] args = new Expression[3 + extraArgs.length];
-    args[0] = Expression.Literal.ofLong(packedPrefixWord(prefix, 0));
-    args[1] = Expression.Literal.ofLong(packedPrefixWord(prefix, Long.BYTES));
-    args[2] = Expression.Literal.ofInt(prefix.length);
-    System.arraycopy(extraArgs, 0, args, 3, extraArgs.length);
-    return args;
-  }
-
-  private static Expression[] objectPackedPrefixArgs(JsonFieldInfo property, Expression value) {
-    byte[] namePrefix = property.utf8NamePrefix();
-    byte[] prefix = new byte[namePrefix.length + 1];
-    prefix[0] = '{';
-    System.arraycopy(namePrefix, 0, prefix, 1, namePrefix.length);
-    return new Expression[] {
-      Expression.Literal.ofLong(packedPrefixWord(prefix, 0)),
-      Expression.Literal.ofLong(packedPrefixWord(prefix, Long.BYTES)),
-      Expression.Literal.ofInt(prefix.length),
-      value
-    };
-  }
-
-  private static Expression[] stringPackedPrefixArgs(
-      JsonFieldInfo property, int id, boolean comma, Expression... extraArgs) {
-    byte[] prefix =
-        comma ? property.stringUtf16CommaNamePrefix() : property.stringUtf16NamePrefix();
-    Expression[] args = new Expression[6 + extraArgs.length];
-    args[0] = fieldRef((comma ? "sc" : "s") + id, byte[].class);
-    args[1] = Expression.Literal.ofLong(packedPrefixWord(prefix, 0));
-    args[2] = Expression.Literal.ofLong(packedPrefixWord(prefix, Long.BYTES));
-    args[3] = Expression.Literal.ofLong(packedPrefixWord(prefix, Long.BYTES * 2));
-    args[4] = Expression.Literal.ofLong(packedPrefixWord(prefix, Long.BYTES * 3));
-    args[5] = Expression.Literal.ofInt(prefix.length);
-    System.arraycopy(extraArgs, 0, args, 6, extraArgs.length);
-    return args;
-  }
-
-  private static Expression[] packedDynamicPrefixArgs(
-      JsonFieldInfo property, Expression index, Expression... extraArgs) {
-    byte[] namePrefix = property.utf8NamePrefix();
-    byte[] commaPrefix = property.utf8CommaNamePrefix();
-    Expression[] args = new Expression[5 + extraArgs.length];
-    args[0] = Expression.Literal.ofLong(packedPrefixWord(namePrefix, 0));
-    args[1] = Expression.Literal.ofLong(packedPrefixWord(commaPrefix, 0));
-    args[2] = Expression.Literal.ofInt(namePrefix.length);
-    args[3] = Expression.Literal.ofInt(commaPrefix.length);
-    args[4] = index;
-    System.arraycopy(extraArgs, 0, args, 5, extraArgs.length);
-    return args;
-  }
-
-  private static boolean canPackPrefix(JsonFieldInfo property, boolean comma) {
-    int length = comma ? property.utf8CommaNamePrefix().length : property.utf8NamePrefix().length;
-    return length <= Long.BYTES * 2;
-  }
-
-  private static boolean canPackStringUtf16Prefix(JsonFieldInfo property, boolean comma) {
-    int length =
-        comma
-            ? property.stringUtf16CommaNamePrefix().length
-            : property.stringUtf16NamePrefix().length;
-    return length <= Long.BYTES * 4;
-  }
-
-  private static boolean canPackSinglePrefix(JsonFieldInfo property, boolean comma) {
-    int length = comma ? property.utf8CommaNamePrefix().length : property.utf8NamePrefix().length;
-    return length <= Long.BYTES;
-  }
-
-  private static long packedPrefixWord(byte[] prefix, int offset) {
+  static long packedPrefixWord(byte[] prefix, int offset) {
     long word = 0;
     int end = Math.min(prefix.length, offset + Long.BYTES);
     for (int i = offset; i < end; i++) {
       word |= (prefix[i] & 0xffL) << ((i - offset) << 3);
     }
     return word;
-  }
-
-  private static Expression stringUtf16EnumFieldValue(
-      int id, Expression value, boolean commaKnown, Expression index) {
-    return new Expression.Invoke(
-            fieldRef("wp" + id, JsonFieldInfo.class),
-            "stringUtf16EnumFieldValue",
-            TypeRef.of(byte[].class),
-            value,
-            commaFlag(commaKnown, index))
-        .inline();
   }
 
   private Expression writeCodec(
@@ -1242,15 +1064,11 @@ abstract class JsonWriterCodegen {
     }
   }
 
-  private static Reference utf16PrefixRef(boolean comma, int id) {
-    return fieldRef((comma ? "sc16" : "s16") + id, byte[].class);
-  }
-
   private static Expression commaFlag(boolean commaKnown, Expression index) {
     return commaKnown ? Expression.Literal.True : ne(index, Expression.Literal.ofInt(0));
   }
 
-  private static Expression increment(Expression value) {
+  static Expression increment(Expression value) {
     return new Expression.Assign(value, add(value, Expression.Literal.ofInt(1)));
   }
 
@@ -1264,533 +1082,7 @@ abstract class JsonWriterCodegen {
         || kind == JsonFieldKind.ENUM;
   }
 
-  static final class StringGenerator extends JsonWriterCodegen {
-    StringGenerator(JsonCodegen codegen) {
-      super(codegen);
-    }
-
-    @Override
-    Class<?> codecFieldType(JsonFieldInfo property) {
-      return codegen.stringWriterFieldType(property.writeTypeInfo());
-    }
-
-    @Override
-    Class<?> writerType() {
-      return StringJsonWriter.class;
-    }
-
-    @Override
-    Class<?> codecArrayType() {
-      return StringWriterCodec[].class;
-    }
-
-    @Override
-    Class<?> objectWriterType() {
-      return StringObjectWriter.class;
-    }
-
-    @Override
-    Class<?> completeWriterType() {
-      return StringWriterCodec.class;
-    }
-
-    @Override
-    String writeMethod() {
-      return "writeString";
-    }
-
-    @Override
-    String membersMethod() {
-      return "writeStringMembers";
-    }
-
-    @Override
-    String writeAnyMethod() {
-      return "writeStringAny";
-    }
-
-    @Override
-    int splitMemberThreshold() {
-      return MIN_STRING_SPLIT_MEMBERS;
-    }
-
-    @Override
-    PrefixFields prefixFields(JsonFieldInfo[] properties, boolean objectStartFused) {
-      return stringPrefixFields(properties, objectStartFused);
-    }
-
-    @Override
-    void addPrefixFields(CodegenContext ctx, JsonFieldInfo property, int id, PrefixFields fields) {
-      ctx.addField(byte[].class, "s" + id);
-      ctx.addField(byte[].class, "sc" + id);
-      if (fields.name[id]) {
-        ctx.addField(byte[].class, "s16" + id);
-      }
-      if (fields.comma[id]) {
-        ctx.addField(byte[].class, "sc16" + id);
-      }
-    }
-
-    @Override
-    void addPrefixAssignments(
-        Expression.ListExpression expressions,
-        Expression property,
-        JsonFieldInfo field,
-        int id,
-        PrefixFields fields) {
-      expressions.add(
-          new Expression.Assign(
-              stringPrefixRef(false, id),
-              new Expression.Invoke(property, "stringNamePrefix", TypeRef.of(byte[].class))
-                  .inline()));
-      expressions.add(
-          new Expression.Assign(
-              stringPrefixRef(true, id),
-              new Expression.Invoke(property, "stringCommaNamePrefix", TypeRef.of(byte[].class))
-                  .inline()));
-      if (fields.name[id]) {
-        expressions.add(
-            new Expression.Assign(
-                utf16PrefixRef(false, id),
-                new Expression.Invoke(property, "stringUtf16NamePrefix", TypeRef.of(byte[].class))
-                    .inline()));
-      }
-      if (fields.comma[id]) {
-        expressions.add(
-            new Expression.Assign(
-                utf16PrefixRef(true, id),
-                new Expression.Invoke(
-                        property, "stringUtf16CommaNamePrefix", TypeRef.of(byte[].class))
-                    .inline()));
-      }
-    }
-
-    @Override
-    Reference writerRef() {
-      return new Reference("writer", TypeRef.of(StringJsonWriter.class));
-    }
-
-    @Override
-    Expression writeObjectStartPrimitive(
-        JsonFieldInfo property, Expression value, Expression writer) {
-      String method;
-      switch (property.writeKind()) {
-        case BYTE:
-        case SHORT:
-        case INT:
-          method = "writeObjectStartWithIntField";
-          break;
-        case LONG:
-          method = "writeObjectStartWithLongField";
-          break;
-        default:
-          throw new ForyJsonException(
-              "Unsupported generated object-start kind " + property.writeKind());
-      }
-      if (canPackStringUtf16Prefix(property, false)) {
-        return new Expression.Invoke(
-            writer, method, stringPackedPrefixArgs(property, 0, false, value));
-      }
-      return new Expression.Invoke(
-          writer, method, stringPrefixRef(false, 0), utf16PrefixRef(false, 0), value);
-    }
-
-    @Override
-    Expression tryWriteObjectStartString(
-        JsonFieldInfo property, Expression value, Expression writer) {
-      return null;
-    }
-
-    @Override
-    Expression writeNumberField(
-        JsonFieldInfo property,
-        int id,
-        Expression value,
-        boolean longValue,
-        boolean commaKnown,
-        Expression index,
-        Expression writer) {
-      String method = longValue ? "writeLongField" : "writeIntField";
-      if (commaKnown) {
-        if (canPackStringUtf16Prefix(property, true)) {
-          return new Expression.Invoke(
-              writer, method, stringPackedPrefixArgs(property, id, true, value));
-        }
-        return new Expression.Invoke(
-            writer, method, stringPrefixRef(true, id), utf16PrefixRef(true, id), value);
-      }
-      Expression.ListExpression expressions =
-          new Expression.ListExpression(
-              new Expression.Invoke(
-                  writer,
-                  method,
-                  stringPrefixRef(false, id),
-                  stringPrefixRef(true, id),
-                  utf16PrefixRef(false, id),
-                  utf16PrefixRef(true, id),
-                  index,
-                  value));
-      expressions.add(increment(index));
-      return expressions;
-    }
-
-    @Override
-    Expression writeStringField(
-        JsonFieldInfo property,
-        int id,
-        Expression value,
-        boolean commaKnown,
-        Expression index,
-        Expression writer) {
-      if (commaKnown) {
-        if (canPackStringUtf16Prefix(property, true)) {
-          return new Expression.Invoke(
-              writer, "writeStringField", stringPackedPrefixArgs(property, id, true, value));
-        }
-        return new Expression.Invoke(
-            writer, "writeStringField", stringPrefixRef(true, id), utf16PrefixRef(true, id), value);
-      }
-      Expression.ListExpression expressions =
-          new Expression.ListExpression(
-              new Expression.Invoke(
-                  writer,
-                  "writeStringField",
-                  stringPrefixRef(false, id),
-                  stringPrefixRef(true, id),
-                  utf16PrefixRef(false, id),
-                  utf16PrefixRef(true, id),
-                  index,
-                  value));
-      expressions.add(increment(index));
-      return expressions;
-    }
-
-    @Override
-    Expression writeFieldName(
-        JsonFieldInfo property, int id, boolean commaKnown, Expression index, Expression writer) {
-      if (commaKnown) {
-        if (canPackStringUtf16Prefix(property, true)) {
-          return new Expression.Invoke(
-              writer, "writeRawValue", stringPackedPrefixArgs(property, id, true));
-        }
-        return new Expression.Invoke(
-            writer, "writeRawValue", stringPrefixRef(true, id), utf16PrefixRef(true, id));
-      }
-      Expression.ListExpression expressions =
-          new Expression.ListExpression(
-              new Expression.Invoke(
-                  writer,
-                  "writeRawValue",
-                  stringPrefixRef(false, id),
-                  stringPrefixRef(true, id),
-                  utf16PrefixRef(false, id),
-                  utf16PrefixRef(true, id),
-                  index));
-      expressions.add(increment(index));
-      return expressions;
-    }
-
-    @Override
-    Expression booleanFieldValue(int id, Expression value, boolean commaKnown, Expression index) {
-      return fieldValue(id, "stringBooleanFieldValue", value, commaKnown, index);
-    }
-
-    @Override
-    Expression enumFieldValue(int id, Expression value, boolean commaKnown, Expression index) {
-      return fieldValue(id, "stringEnumFieldValue", value, commaKnown, index);
-    }
-
-    @Override
-    Expression utf16EnumFieldValue(int id, Expression value, boolean commaKnown, Expression index) {
-      return stringUtf16EnumFieldValue(id, value, commaKnown, index);
-    }
-
-    @Override
-    Expression writeExactScalar(JsonFieldInfo property, Expression value, Expression writer) {
-      return null;
-    }
-
-    @Override
-    Expression writeExactArray(JsonFieldInfo property, Expression value, Expression writer) {
-      return null;
-    }
-
-    private static Reference stringPrefixRef(boolean comma, int id) {
-      return fieldRef((comma ? "sc" : "s") + id, byte[].class);
-    }
-  }
-
-  static final class Utf8Generator extends JsonWriterCodegen {
-    Utf8Generator(JsonCodegen codegen) {
-      super(codegen);
-    }
-
-    @Override
-    Class<?> codecFieldType(JsonFieldInfo property) {
-      return codegen.utf8WriterFieldType(property.writeTypeInfo());
-    }
-
-    @Override
-    Class<?> writerType() {
-      return Utf8JsonWriter.class;
-    }
-
-    @Override
-    Class<?> codecArrayType() {
-      return Utf8WriterCodec[].class;
-    }
-
-    @Override
-    Class<?> objectWriterType() {
-      return Utf8ObjectWriter.class;
-    }
-
-    @Override
-    Class<?> completeWriterType() {
-      return Utf8WriterCodec.class;
-    }
-
-    @Override
-    String writeMethod() {
-      return "writeUtf8";
-    }
-
-    @Override
-    String membersMethod() {
-      return "writeUtf8Members";
-    }
-
-    @Override
-    String writeAnyMethod() {
-      return "writeUtf8Any";
-    }
-
-    @Override
-    int splitMemberThreshold() {
-      return MIN_UTF8_SPLIT_MEMBERS;
-    }
-
-    @Override
-    PrefixFields prefixFields(JsonFieldInfo[] properties, boolean objectStartFused) {
-      return utf8PrefixFields(properties, objectStartFused);
-    }
-
-    @Override
-    void addPrefixFields(CodegenContext ctx, JsonFieldInfo property, int id, PrefixFields fields) {
-      if (fields.name[id]) {
-        ctx.addField(byte[].class, "u" + id);
-      }
-      if (fields.comma[id]) {
-        ctx.addField(byte[].class, "uc" + id);
-      }
-    }
-
-    @Override
-    void addPrefixAssignments(
-        Expression.ListExpression expressions,
-        Expression property,
-        JsonFieldInfo field,
-        int id,
-        PrefixFields fields) {
-      if (fields.name[id]) {
-        expressions.add(
-            new Expression.Assign(
-                utf8PrefixRef(false, id),
-                new Expression.Invoke(property, "utf8NamePrefix", TypeRef.of(byte[].class))
-                    .inline()));
-      }
-      if (fields.comma[id]) {
-        expressions.add(
-            new Expression.Assign(
-                utf8PrefixRef(true, id),
-                new Expression.Invoke(property, "utf8CommaNamePrefix", TypeRef.of(byte[].class))
-                    .inline()));
-      }
-    }
-
-    @Override
-    Reference writerRef() {
-      return new Reference("writer", TypeRef.of(Utf8JsonWriter.class));
-    }
-
-    @Override
-    Expression writeObjectStartPrimitive(
-        JsonFieldInfo property, Expression value, Expression writer) {
-      String method;
-      switch (property.writeKind()) {
-        case BYTE:
-        case SHORT:
-        case INT:
-          method = "writeObjectStartWithIntField";
-          break;
-        case LONG:
-          method = "writeObjectStartWithLongField";
-          break;
-        default:
-          throw new ForyJsonException(
-              "Unsupported generated object-start kind " + property.writeKind());
-      }
-      if (canPackPrefix(property, false)) {
-        return new Expression.Invoke(writer, method, packedPrefixArgs(property, false, value));
-      }
-      return new Expression.Invoke(writer, method, utf8PrefixRef(false, 0), value);
-    }
-
-    @Override
-    Expression tryWriteObjectStartString(
-        JsonFieldInfo property, Expression value, Expression writer) {
-      if (!canPackObjectStartString(property)) {
-        return null;
-      }
-      return new Expression.Invoke(
-          writer, "writeObjectStartWithStringField", objectPackedPrefixArgs(property, value));
-    }
-
-    private static boolean canPackObjectStartString(JsonFieldInfo property) {
-      return property.writeKind() == JsonFieldKind.STRING
-          && property.utf8NamePrefix().length < Long.BYTES * 2;
-    }
-
-    @Override
-    Expression writeNumberField(
-        JsonFieldInfo property,
-        int id,
-        Expression value,
-        boolean longValue,
-        boolean commaKnown,
-        Expression index,
-        Expression writer) {
-      String method = longValue ? "writeLongField" : "writeIntField";
-      if (commaKnown) {
-        if (canPackPrefix(property, true)) {
-          return new Expression.Invoke(writer, method, packedPrefixArgs(property, true, value));
-        }
-        return new Expression.Invoke(writer, method, utf8PrefixRef(true, id), value);
-      }
-      if (canPackSinglePrefix(property, false) && canPackSinglePrefix(property, true)) {
-        return new Expression.ListExpression(
-            new Expression.Invoke(writer, method, packedDynamicPrefixArgs(property, index, value)),
-            increment(index));
-      }
-      Expression.ListExpression expressions =
-          new Expression.ListExpression(
-              new Expression.Invoke(
-                  writer, method, utf8PrefixRef(false, id), utf8PrefixRef(true, id), index, value));
-      expressions.add(increment(index));
-      return expressions;
-    }
-
-    @Override
-    Expression writeStringField(
-        JsonFieldInfo property,
-        int id,
-        Expression value,
-        boolean commaKnown,
-        Expression index,
-        Expression writer) {
-      if (commaKnown) {
-        if (canPackPrefix(property, true)) {
-          return new Expression.Invoke(
-              writer, "writeStringField", packedPrefixArgs(property, true, value));
-        }
-        return new Expression.Invoke(writer, "writeStringField", utf8PrefixRef(true, id), value);
-      }
-      if (canPackSinglePrefix(property, false) && canPackSinglePrefix(property, true)) {
-        return new Expression.ListExpression(
-            new Expression.Invoke(
-                writer, "writeStringField", packedDynamicPrefixArgs(property, index, value)),
-            increment(index));
-      }
-      Expression.ListExpression expressions =
-          new Expression.ListExpression(
-              new Expression.Invoke(
-                  writer,
-                  "writeStringField",
-                  utf8PrefixRef(false, id),
-                  utf8PrefixRef(true, id),
-                  index,
-                  value));
-      expressions.add(increment(index));
-      return expressions;
-    }
-
-    @Override
-    Expression writeFieldName(
-        JsonFieldInfo property, int id, boolean commaKnown, Expression index, Expression writer) {
-      if (commaKnown && canPackPrefix(property, true)) {
-        return new Expression.Invoke(writer, "writeRawValue", packedPrefixArgs(property, true));
-      }
-      Expression prefix =
-          commaKnown
-              ? utf8PrefixRef(true, id)
-              : new Expression.Ternary(
-                  eq(index, Expression.Literal.ofInt(0)),
-                  utf8PrefixRef(false, id),
-                  utf8PrefixRef(true, id),
-                  true,
-                  TypeRef.of(byte[].class));
-      Expression.ListExpression expressions =
-          new Expression.ListExpression(new Expression.Invoke(writer, "writeRawValue", prefix));
-      if (!commaKnown) {
-        expressions.add(increment(index));
-      }
-      return expressions;
-    }
-
-    @Override
-    Expression booleanFieldValue(int id, Expression value, boolean commaKnown, Expression index) {
-      return fieldValue(id, "utf8BooleanFieldValue", value, commaKnown, index);
-    }
-
-    @Override
-    Expression enumFieldValue(int id, Expression value, boolean commaKnown, Expression index) {
-      return fieldValue(id, "utf8EnumFieldValue", value, commaKnown, index);
-    }
-
-    @Override
-    Expression utf16EnumFieldValue(int id, Expression value, boolean commaKnown, Expression index) {
-      return null;
-    }
-
-    @Override
-    Expression writeExactScalar(JsonFieldInfo property, Expression value, Expression writer) {
-      Class<?> rawType = property.writeRawType();
-      Object codec = property.writeTypeInfo().utf8Writer();
-      String method;
-      if (rawType == UUID.class && codec == ScalarCodecs.UuidCodec.INSTANCE) {
-        method = "writeUuid";
-      } else if (rawType == LocalDate.class && codec == ScalarCodecs.LocalDateCodec.INSTANCE) {
-        method = "writeLocalDate";
-      } else if (rawType == OffsetDateTime.class
-          && codec == ScalarCodecs.OffsetDateTimeCodec.INSTANCE) {
-        method = "writeOffsetDateTime";
-      } else if (rawType == BigDecimal.class && codec == ScalarCodecs.BigDecimalCodec.INSTANCE) {
-        method = "writeBigDecimal";
-      } else {
-        return null;
-      }
-      return new Expression.Invoke(writer, method, value);
-    }
-
-    @Override
-    Expression writeExactArray(JsonFieldInfo property, Expression value, Expression writer) {
-      Class<?> rawType = property.writeRawType();
-      Class<?> codecType = property.writeTypeInfo().utf8Writer().getClass();
-      if (rawType == String[].class && codecType == ArrayCodec.StringArrayCodec.class) {
-        return new Expression.Invoke(writer, "writeStringArray", value);
-      }
-      if (rawType == long[].class && codecType == ArrayCodec.LongArrayCodec.class) {
-        return new Expression.Invoke(writer, "writeLongArray", value);
-      }
-      return null;
-    }
-
-    private static Reference utf8PrefixRef(boolean comma, int id) {
-      return fieldRef((comma ? "uc" : "u") + id, byte[].class);
-    }
-  }
-
-  private static Expression fieldValue(
+  static Expression fieldValue(
       int id, String method, Expression value, boolean commaKnown, Expression index) {
     return new Expression.Invoke(
             fieldRef("wp" + id, JsonFieldInfo.class),
