+++ java/fory-json/src/main/java/org/apache/fory/json/resolver/JsonTypeResolver.java
@@ -19,9 +19,6 @@
 
 package org.apache.fory.json.resolver;
 
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodType;
-import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Type;
 import java.util.HashMap;
@@ -31,7 +28,6 @@ import java.util.Map;
 import java.util.Set;
 import org.apache.fory.annotation.Internal;
 import org.apache.fory.collection.IdentityMap;
-import org.apache.fory.json.ForyJsonException;
 import org.apache.fory.json.codec.ClosedSubtypeCodec;
 import org.apache.fory.json.codec.CodecUtils;
 import org.apache.fory.json.codec.JsonCodec;
@@ -51,8 +47,6 @@ import org.apache.fory.json.meta.JsonCreatorFieldInfo;
 import org.apache.fory.json.meta.JsonCreatorInfo;
 import org.apache.fory.json.meta.JsonFieldInfo;
 import org.apache.fory.json.meta.JsonFieldTable;
-import org.apache.fory.platform.AndroidSupport;
-import org.apache.fory.platform.internal._JDKAccess;
 import org.apache.fory.reflect.ReflectionUtils;
 import org.apache.fory.reflect.TypeRef;
 
@@ -579,10 +573,11 @@ public final class JsonTypeResolver {
     }
     AnyInfo any = owner.anyInfo();
     if (any == null || any.writeField() == null && any.writeGetter() == null) {
-      return instantiateStringWriter(generatedClass, fields, codecs);
+      return GeneratedCodecInstantiator.instantiateStringWriter(generatedClass, fields, codecs);
     }
     if (!storesAnyCodec(owner, any)) {
-      return instantiateAnyStringWriter(generatedClass, owner, fields, codecs);
+      return GeneratedCodecInstantiator.instantiateAnyStringWriter(
+          generatedClass, owner, fields, codecs);
     }
     StringWriterCodec<Object> anyCodec = any.valueTypeInfo().stringWriter();
     if (any.valueTypeInfo().usesDefaultObjectCodec()
@@ -590,7 +585,8 @@ public final class JsonTypeResolver {
         && anyCodec instanceof ObjectCodec) {
       anyCodec = stringWriter((ObjectCodec<Object>) anyCodec);
     }
-    return instantiateAnyStringWriter(generatedClass, owner, fields, codecs, anyCodec);
+    return GeneratedCodecInstantiator.instantiateAnyStringWriter(
+        generatedClass, owner, fields, codecs, anyCodec);
   }
 
   @SuppressWarnings("unchecked")
@@ -613,10 +609,11 @@ public final class JsonTypeResolver {
     }
     AnyInfo any = owner.anyInfo();
     if (any == null || any.writeField() == null && any.writeGetter() == null) {
-      return instantiateUtf8Writer(generatedClass, fields, codecs);
+      return GeneratedCodecInstantiator.instantiateUtf8Writer(generatedClass, fields, codecs);
     }
     if (!storesAnyCodec(owner, any)) {
-      return instantiateAnyUtf8Writer(generatedClass, owner, fields, codecs);
+      return GeneratedCodecInstantiator.instantiateAnyUtf8Writer(
+          generatedClass, owner, fields, codecs);
     }
     Utf8WriterCodec<Object> anyCodec = any.valueTypeInfo().utf8Writer();
     if (any.valueTypeInfo().usesDefaultObjectCodec()
@@ -624,7 +621,8 @@ public final class JsonTypeResolver {
         && anyCodec instanceof ObjectCodec) {
       anyCodec = utf8Writer((ObjectCodec<Object>) anyCodec);
     }
-    return instantiateAnyUtf8Writer(generatedClass, owner, fields, codecs, anyCodec);
+    return GeneratedCodecInstantiator.instantiateAnyUtf8Writer(
+        generatedClass, owner, fields, codecs, anyCodec);
   }
 
   @SuppressWarnings("unchecked")
@@ -646,10 +644,12 @@ public final class JsonTypeResolver {
       }
       AnyInfo any = owner.anyInfo();
       if (any == null || any.readField() == null && any.readSetter() == null) {
-        return instantiateLatin1Reader(generatedClass, owner, fields, codecs);
+        return GeneratedCodecInstantiator.instantiateLatin1Reader(
+            generatedClass, owner, fields, codecs);
       }
       if (!storesAnyCodec(owner, any)) {
-        return instantiateAnyLatin1Reader(generatedClass, owner, readTable, fields, codecs);
+        return GeneratedCodecInstantiator.instantiateAnyLatin1Reader(
+            generatedClass, owner, readTable, fields, codecs);
       }
       Latin1ReaderCodec<Object> anyCodec = any.valueTypeInfo().latin1Reader();
       if (any.valueTypeInfo().usesDefaultObjectCodec()
@@ -657,7 +657,8 @@ public final class JsonTypeResolver {
           && anyCodec instanceof ObjectCodec) {
         anyCodec = latin1Reader((ObjectCodec<Object>) anyCodec);
       }
-      return instantiateAnyLatin1Reader(generatedClass, owner, readTable, fields, codecs, anyCodec);
+      return GeneratedCodecInstantiator.instantiateAnyLatin1Reader(
+          generatedClass, owner, readTable, fields, codecs, anyCodec);
     }
     Latin1ReaderCodec<Object>[] codecs =
         (Latin1ReaderCodec<Object>[]) new Latin1ReaderCodec<?>[fields.length];
@@ -674,10 +675,12 @@ public final class JsonTypeResolver {
     }
     AnyInfo any = owner.anyInfo();
     if (any == null || any.readField() == null && any.readSetter() == null) {
-      return instantiateLatin1Reader(generatedClass, owner, fields, codecs);
+      return GeneratedCodecInstantiator.instantiateLatin1Reader(
+          generatedClass, owner, fields, codecs);
     }
     if (!storesAnyCodec(owner, any)) {
-      return instantiateAnyLatin1Reader(generatedClass, owner, readTable, fields, codecs);
+      return GeneratedCodecInstantiator.instantiateAnyLatin1Reader(
+          generatedClass, owner, readTable, fields, codecs);
     }
     Latin1ReaderCodec<Object> anyCodec = any.valueTypeInfo().latin1Reader();
     if (any.valueTypeInfo().usesDefaultObjectCodec()
@@ -685,7 +688,8 @@ public final class JsonTypeResolver {
         && anyCodec instanceof ObjectCodec) {
       anyCodec = latin1Reader((ObjectCodec<Object>) anyCodec);
     }
-    return instantiateAnyLatin1Reader(generatedClass, owner, readTable, fields, codecs, anyCodec);
+    return GeneratedCodecInstantiator.instantiateAnyLatin1Reader(
+        generatedClass, owner, readTable, fields, codecs, anyCodec);
   }
 
   @SuppressWarnings("unchecked")
@@ -707,10 +711,12 @@ public final class JsonTypeResolver {
       }
       AnyInfo any = owner.anyInfo();
       if (any == null || any.readField() == null && any.readSetter() == null) {
-        return instantiateUtf16Reader(generatedClass, owner, fields, codecs);
+        return GeneratedCodecInstantiator.instantiateUtf16Reader(
+            generatedClass, owner, fields, codecs);
       }
       if (!storesAnyCodec(owner, any)) {
-        return instantiateAnyUtf16Reader(generatedClass, owner, readTable, fields, codecs);
+        return GeneratedCodecInstantiator.instantiateAnyUtf16Reader(
+            generatedClass, owner, readTable, fields, codecs);
       }
       Utf16ReaderCodec<Object> anyCodec = any.valueTypeInfo().utf16Reader();
       if (any.valueTypeInfo().usesDefaultObjectCodec()
@@ -718,7 +724,8 @@ public final class JsonTypeResolver {
           && anyCodec instanceof ObjectCodec) {
         anyCodec = utf16Reader((ObjectCodec<Object>) anyCodec);
       }
-      return instantiateAnyUtf16Reader(generatedClass, owner, readTable, fields, codecs, anyCodec);
+      return GeneratedCodecInstantiator.instantiateAnyUtf16Reader(
+          generatedClass, owner, readTable, fields, codecs, anyCodec);
     }
     Utf16ReaderCodec<Object>[] codecs =
         (Utf16ReaderCodec<Object>[]) new Utf16ReaderCodec<?>[fields.length];
@@ -734,10 +741,12 @@ public final class JsonTypeResolver {
     }
     AnyInfo any = owner.anyInfo();
     if (any == null || any.readField() == null && any.readSetter() == null) {
-      return instantiateUtf16Reader(generatedClass, owner, fields, codecs);
+      return GeneratedCodecInstantiator.instantiateUtf16Reader(
+          generatedClass, owner, fields, codecs);
     }
     if (!storesAnyCodec(owner, any)) {
-      return instantiateAnyUtf16Reader(generatedClass, owner, readTable, fields, codecs);
+      return GeneratedCodecInstantiator.instantiateAnyUtf16Reader(
+          generatedClass, owner, readTable, fields, codecs);
     }
     Utf16ReaderCodec<Object> anyCodec = any.valueTypeInfo().utf16Reader();
     if (any.valueTypeInfo().usesDefaultObjectCodec()
@@ -745,7 +754,8 @@ public final class JsonTypeResolver {
         && anyCodec instanceof ObjectCodec) {
       anyCodec = utf16Reader((ObjectCodec<Object>) anyCodec);
     }
-    return instantiateAnyUtf16Reader(generatedClass, owner, readTable, fields, codecs, anyCodec);
+    return GeneratedCodecInstantiator.instantiateAnyUtf16Reader(
+        generatedClass, owner, readTable, fields, codecs, anyCodec);
   }
 
   @SuppressWarnings("unchecked")
@@ -767,10 +777,12 @@ public final class JsonTypeResolver {
       }
       AnyInfo any = owner.anyInfo();
       if (any == null || any.readField() == null && any.readSetter() == null) {
-        return instantiateUtf8Reader(generatedClass, owner, fields, codecs);
+        return GeneratedCodecInstantiator.instantiateUtf8Reader(
+            generatedClass, owner, fields, codecs);
       }
       if (!storesAnyCodec(owner, any)) {
-        return instantiateAnyUtf8Reader(generatedClass, owner, readTable, fields, codecs);
+        return GeneratedCodecInstantiator.instantiateAnyUtf8Reader(
+            generatedClass, owner, readTable, fields, codecs);
       }
       Utf8ReaderCodec<Object> anyCodec = any.valueTypeInfo().utf8Reader();
       if (any.valueTypeInfo().usesDefaultObjectCodec()
@@ -778,7 +790,8 @@ public final class JsonTypeResolver {
           && anyCodec instanceof ObjectCodec) {
         anyCodec = utf8Reader((ObjectCodec<Object>) anyCodec);
       }
-      return instantiateAnyUtf8Reader(generatedClass, owner, readTable, fields, codecs, anyCodec);
+      return GeneratedCodecInstantiator.instantiateAnyUtf8Reader(
+          generatedClass, owner, readTable, fields, codecs, anyCodec);
     }
     Utf8ReaderCodec<Object>[] codecs =
         (Utf8ReaderCodec<Object>[]) new Utf8ReaderCodec<?>[fields.length];
@@ -794,10 +807,12 @@ public final class JsonTypeResolver {
     }
     AnyInfo any = owner.anyInfo();
     if (any == null || any.readField() == null && any.readSetter() == null) {
-      return instantiateUtf8Reader(generatedClass, owner, fields, codecs);
+      return GeneratedCodecInstantiator.instantiateUtf8Reader(
+          generatedClass, owner, fields, codecs);
     }
     if (!storesAnyCodec(owner, any)) {
-      return instantiateAnyUtf8Reader(generatedClass, owner, readTable, fields, codecs);
+      return GeneratedCodecInstantiator.instantiateAnyUtf8Reader(
+          generatedClass, owner, readTable, fields, codecs);
     }
     Utf8ReaderCodec<Object> anyCodec = any.valueTypeInfo().utf8Reader();
     if (any.valueTypeInfo().usesDefaultObjectCodec()
@@ -805,7 +820,8 @@ public final class JsonTypeResolver {
         && anyCodec instanceof ObjectCodec) {
       anyCodec = utf8Reader((ObjectCodec<Object>) anyCodec);
     }
-    return instantiateAnyUtf8Reader(generatedClass, owner, readTable, fields, codecs, anyCodec);
+    return GeneratedCodecInstantiator.instantiateAnyUtf8Reader(
+        generatedClass, owner, readTable, fields, codecs, anyCodec);
   }
 
   private Latin1ReaderCodec<Object> newInlineLatin1Reader(
@@ -1341,483 +1357,6 @@ public final class JsonTypeResolver {
     if (codec.getClass() != result && !refinement.isInstance(codec)) {
       throw new IllegalStateException(
           "Generated JSON callback does not match installed writer capability");
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static StringWriterCodec<Object> instantiateStringWriter(
-      Class<?> type, JsonFieldInfo[] fields, StringWriterCodec<Object>[] codecs) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(JsonFieldInfo[].class, StringWriterCodec[].class);
-        constructor.setAccessible(true);
-        return (StringWriterCodec<Object>) constructor.newInstance(fields, codecs);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class, JsonFieldInfo[].class, StringWriterCodec[].class));
-      return (StringWriterCodec<Object>) constructor.invoke(fields, codecs);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON String writer", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static StringWriterCodec<Object> instantiateAnyStringWriter(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldInfo[] fields,
-      StringWriterCodec<Object>[] codecs) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class, JsonFieldInfo[].class, StringWriterCodec[].class);
-        constructor.setAccessible(true);
-        return (StringWriterCodec<Object>) constructor.newInstance(owner, fields, codecs);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldInfo[].class,
-                      StringWriterCodec[].class));
-      return (StringWriterCodec<Object>) constructor.invoke(owner, fields, codecs);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON Any String writer", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static StringWriterCodec<Object> instantiateAnyStringWriter(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldInfo[] fields,
-      StringWriterCodec<Object>[] codecs,
-      StringWriterCodec<Object> anyCodec) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class,
-                JsonFieldInfo[].class,
-                StringWriterCodec[].class,
-                StringWriterCodec.class);
-        constructor.setAccessible(true);
-        return (StringWriterCodec<Object>) constructor.newInstance(owner, fields, codecs, anyCodec);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldInfo[].class,
-                      StringWriterCodec[].class,
-                      StringWriterCodec.class));
-      return (StringWriterCodec<Object>) constructor.invoke(owner, fields, codecs, anyCodec);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON Any String writer", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Utf8WriterCodec<Object> instantiateUtf8Writer(
-      Class<?> type, JsonFieldInfo[] fields, Utf8WriterCodec<Object>[] codecs) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(JsonFieldInfo[].class, Utf8WriterCodec[].class);
-        constructor.setAccessible(true);
-        return (Utf8WriterCodec<Object>) constructor.newInstance(fields, codecs);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class, JsonFieldInfo[].class, Utf8WriterCodec[].class));
-      return (Utf8WriterCodec<Object>) constructor.invoke(fields, codecs);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON UTF8 writer", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Utf8WriterCodec<Object> instantiateAnyUtf8Writer(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldInfo[] fields,
-      Utf8WriterCodec<Object>[] codecs) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class, JsonFieldInfo[].class, Utf8WriterCodec[].class);
-        constructor.setAccessible(true);
-        return (Utf8WriterCodec<Object>) constructor.newInstance(owner, fields, codecs);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldInfo[].class,
-                      Utf8WriterCodec[].class));
-      return (Utf8WriterCodec<Object>) constructor.invoke(owner, fields, codecs);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON Any UTF8 writer", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Utf8WriterCodec<Object> instantiateAnyUtf8Writer(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldInfo[] fields,
-      Utf8WriterCodec<Object>[] codecs,
-      Utf8WriterCodec<Object> anyCodec) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class,
-                JsonFieldInfo[].class,
-                Utf8WriterCodec[].class,
-                Utf8WriterCodec.class);
-        constructor.setAccessible(true);
-        return (Utf8WriterCodec<Object>) constructor.newInstance(owner, fields, codecs, anyCodec);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldInfo[].class,
-                      Utf8WriterCodec[].class,
-                      Utf8WriterCodec.class));
-      return (Utf8WriterCodec<Object>) constructor.invoke(owner, fields, codecs, anyCodec);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON Any UTF8 writer", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Latin1ReaderCodec<Object> instantiateLatin1Reader(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldInfo[] fields,
-      Latin1ReaderCodec<Object>[] codecs) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class, JsonFieldInfo[].class, Latin1ReaderCodec[].class);
-        constructor.setAccessible(true);
-        return (Latin1ReaderCodec<Object>) constructor.newInstance(owner, fields, codecs);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldInfo[].class,
-                      Latin1ReaderCodec[].class));
-      return (Latin1ReaderCodec<Object>) constructor.invoke(owner, fields, codecs);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON Latin1 reader", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Latin1ReaderCodec<Object> instantiateAnyLatin1Reader(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldTable readTable,
-      JsonFieldInfo[] fields,
-      Latin1ReaderCodec<Object>[] codecs) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class,
-                JsonFieldTable.class,
-                JsonFieldInfo[].class,
-                Latin1ReaderCodec[].class);
-        constructor.setAccessible(true);
-        return (Latin1ReaderCodec<Object>)
-            constructor.newInstance(owner, readTable, fields, codecs);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldTable.class,
-                      JsonFieldInfo[].class,
-                      Latin1ReaderCodec[].class));
-      return (Latin1ReaderCodec<Object>) constructor.invoke(owner, readTable, fields, codecs);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON Any Latin1 reader", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Latin1ReaderCodec<Object> instantiateAnyLatin1Reader(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldTable readTable,
-      JsonFieldInfo[] fields,
-      Latin1ReaderCodec<Object>[] codecs,
-      Latin1ReaderCodec<Object> anyCodec) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class,
-                JsonFieldTable.class,
-                JsonFieldInfo[].class,
-                Latin1ReaderCodec[].class,
-                Latin1ReaderCodec.class);
-        constructor.setAccessible(true);
-        return (Latin1ReaderCodec<Object>)
-            constructor.newInstance(owner, readTable, fields, codecs, anyCodec);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldTable.class,
-                      JsonFieldInfo[].class,
-                      Latin1ReaderCodec[].class,
-                      Latin1ReaderCodec.class));
-      return (Latin1ReaderCodec<Object>)
-          constructor.invoke(owner, readTable, fields, codecs, anyCodec);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON Any Latin1 reader", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Utf16ReaderCodec<Object> instantiateUtf16Reader(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldInfo[] fields,
-      Utf16ReaderCodec<Object>[] codecs) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class, JsonFieldInfo[].class, Utf16ReaderCodec[].class);
-        constructor.setAccessible(true);
-        return (Utf16ReaderCodec<Object>) constructor.newInstance(owner, fields, codecs);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldInfo[].class,
-                      Utf16ReaderCodec[].class));
-      return (Utf16ReaderCodec<Object>) constructor.invoke(owner, fields, codecs);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON UTF16 reader", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Utf16ReaderCodec<Object> instantiateAnyUtf16Reader(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldTable readTable,
-      JsonFieldInfo[] fields,
-      Utf16ReaderCodec<Object>[] codecs) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class,
-                JsonFieldTable.class,
-                JsonFieldInfo[].class,
-                Utf16ReaderCodec[].class);
-        constructor.setAccessible(true);
-        return (Utf16ReaderCodec<Object>) constructor.newInstance(owner, readTable, fields, codecs);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldTable.class,
-                      JsonFieldInfo[].class,
-                      Utf16ReaderCodec[].class));
-      return (Utf16ReaderCodec<Object>) constructor.invoke(owner, readTable, fields, codecs);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON Any UTF16 reader", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Utf16ReaderCodec<Object> instantiateAnyUtf16Reader(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldTable readTable,
-      JsonFieldInfo[] fields,
-      Utf16ReaderCodec<Object>[] codecs,
-      Utf16ReaderCodec<Object> anyCodec) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class,
-                JsonFieldTable.class,
-                JsonFieldInfo[].class,
-                Utf16ReaderCodec[].class,
-                Utf16ReaderCodec.class);
-        constructor.setAccessible(true);
-        return (Utf16ReaderCodec<Object>)
-            constructor.newInstance(owner, readTable, fields, codecs, anyCodec);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldTable.class,
-                      JsonFieldInfo[].class,
-                      Utf16ReaderCodec[].class,
-                      Utf16ReaderCodec.class));
-      return (Utf16ReaderCodec<Object>)
-          constructor.invoke(owner, readTable, fields, codecs, anyCodec);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON Any UTF16 reader", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Utf8ReaderCodec<Object> instantiateUtf8Reader(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldInfo[] fields,
-      Utf8ReaderCodec<Object>[] codecs) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class, JsonFieldInfo[].class, Utf8ReaderCodec[].class);
-        constructor.setAccessible(true);
-        return (Utf8ReaderCodec<Object>) constructor.newInstance(owner, fields, codecs);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldInfo[].class,
-                      Utf8ReaderCodec[].class));
-      return (Utf8ReaderCodec<Object>) constructor.invoke(owner, fields, codecs);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON UTF8 reader", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Utf8ReaderCodec<Object> instantiateAnyUtf8Reader(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldTable readTable,
-      JsonFieldInfo[] fields,
-      Utf8ReaderCodec<Object>[] codecs) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class,
-                JsonFieldTable.class,
-                JsonFieldInfo[].class,
-                Utf8ReaderCodec[].class);
-        constructor.setAccessible(true);
-        return (Utf8ReaderCodec<Object>) constructor.newInstance(owner, readTable, fields, codecs);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldTable.class,
-                      JsonFieldInfo[].class,
-                      Utf8ReaderCodec[].class));
-      return (Utf8ReaderCodec<Object>) constructor.invoke(owner, readTable, fields, codecs);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON Any UTF8 reader", e);
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Utf8ReaderCodec<Object> instantiateAnyUtf8Reader(
-      Class<?> type,
-      ObjectCodec<?> owner,
-      JsonFieldTable readTable,
-      JsonFieldInfo[] fields,
-      Utf8ReaderCodec<Object>[] codecs,
-      Utf8ReaderCodec<Object> anyCodec) {
-    try {
-      if (AndroidSupport.IS_ANDROID) {
-        Constructor<?> constructor =
-            type.getDeclaredConstructor(
-                ObjectCodec.class,
-                JsonFieldTable.class,
-                JsonFieldInfo[].class,
-                Utf8ReaderCodec[].class,
-                Utf8ReaderCodec.class);
-        constructor.setAccessible(true);
-        return (Utf8ReaderCodec<Object>)
-            constructor.newInstance(owner, readTable, fields, codecs, anyCodec);
-      }
-      MethodHandle constructor =
-          _JDKAccess._trustedLookup(type)
-              .findConstructor(
-                  type,
-                  MethodType.methodType(
-                      void.class,
-                      ObjectCodec.class,
-                      JsonFieldTable.class,
-                      JsonFieldInfo[].class,
-                      Utf8ReaderCodec[].class,
-                      Utf8ReaderCodec.class));
-      return (Utf8ReaderCodec<Object>)
-          constructor.invoke(owner, readTable, fields, codecs, anyCodec);
-    } catch (Throwable e) {
-      throw new ForyJsonException("Cannot instantiate generated JSON Any UTF8 reader", e);
     }
   }
